requirements.txt
Flask
spacy
PyMuPDF
google-generativeai
python-dotenv